Fix issue with resetting point weights when returning to step 2 from step 3
Previously, point weights were being reset when returning from step 3 to step 2 due to a bug in st_canvas handling the initial_drawing=st.session_state.canvas_data. As a temporary solution, coordinates are no longer recreated on this render. However, there is still an issue where moving a point on step 2 causes its weight to be reset, as the point is treated as a new one. To fix this, points should be assigned unique IDs, and changes should be tracked based on these IDs.